requirements.txt
playwright==1.45.0
pytest>=7.0.0,<8.0.0
pytest-playwright==0.4.3
python-dotenv==1.0.0
requests==2.31.0
# Web框架
flask>=2.0.0
flask-socketio>=5.0.0
flask-cors>=4.0.0
python-socketio>=5.0.0
# 数据库支持
flask-sqlalchemy>=3.0.0
# SQLite数据库 (内置支持，无需额外依赖)
# 开发工具
black>=22.0.0
flake8>=4.0.0
pytest-cov>=3.0.0
# 测试工具
factory-boy>=3.2.0
faker>=18.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0